Fixes #260: Refactor actions.js file
